+++ cookbooks/chub_albers/metadata.rb
@@ -1,9 +1,8 @@
 name             'chub_albers'
-maintainer       'CommerceHub Inc.'
-maintainer_email 'lzarou@commercehub.com'
-license          'Apache 2.0'
+maintainer       'Commerce Technology, Inc.'
+maintainer_email 'dcross@commercehub.com'
+license          'All rights reserved'
 description      'Installs/Configures chub_albers'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'
-
-depends 'java'
+depends          'windows'