+++ cookbooks/chub_windows/attributes/default.rb
@@ -1,0 +1,6 @@
+#
+# Author:: David Cross (dcross@commercehub.com)
+# Cookbook Name:: chub_windows
+# Copyright 2014, Commerce Technology, Inc.
+
+default['windows']['java']['home'] = 'c:\Program Files\Java