+++ cookbooks/chub_windows/metadata.rb
@@ -1,0 +1,9 @@
+name             'chub_windows'
+maintainer       'Commerce Technology, Inc.'
+maintainer_email 'dcross@commercehub.com'
+license          'All rights reserved'
+description      'Installs/Configures windows for chub base'
+long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
+version          '0.1.0'
+depends          'windows'
+depends          'java'