+++ cookbooks/chub-klerk/CHANGELOG.md
@@ -2,6 +2,12 @@ chub-klerk Cookbook CHANGELOG
 =======================
 This file is used to list changes made in each version of the chub-klerk cookbook.
 
+v0.10.5
+-------
+- Update java cookbook version constraint to 1.22.0
+- Update jdk_version to 8
+- Remove support for specification of JVM permanent generation size as it is no longer applicable as of Java 8
+
 v0.10.4
 -------
 - Add support for specification of JVM permanent generation size via attribute `java_perm_gen_size`, with a default value of `128m`